+++ source/common/console/c_console.h
@@ -57,7 +57,7 @@ extern uint8_t ConsoleState;
 // Initialize the console
 void C_InitConsole (int width, int height, bool ingame);
 void C_DeinitConsole ();
-void C_InitConback(FTextureID fallback, bool tile, double lightlevel = 1.);
+void C_InitConback(FTextureID fallback, bool tile, float lightlevel = 1.);
 
 // Adjust the console for a new screen mode
 void C_NewModeAdjust (void);
@@ -81,8 +81,8 @@ void C_SetNotifyBuffer(FNotifyBufferBase *nbb);
 
 bool C_Responder (event_t *ev);
 
-extern double NotifyFontScale;
-void C_SetNotifyFontScale(double scale);
+extern float NotifyFontScale;
+void C_SetNotifyFontScale(float scale);
 
 extern const char *console_bar;
 extern int chatmodeon;
